+++ linphone-app/ui/views/App/Main/MainWindowMenuBar.qml
@@ -18,6 +18,8 @@ Item {
 		menu.close()
 	}
 	
+	signal displayRecordings()
+	
 	// ---------------------------------------------------------------------------
 	// Shortcuts.
 	// ---------------------------------------------------------------------------
@@ -50,6 +52,11 @@ Item {
 		}
 	}
 	
+	Shortcut {
+		id: recordingsShortcut
+		onActivated: menuParent.displayRecordings()
+	}
+	
 	// ---------------------------------------------------------------------------
 	// Menu.
 	// ---------------------------------------------------------------------------
@@ -62,6 +69,12 @@ Item {
 			text: qsTr('settings')
 			
 			onTriggered: settingsShortcut.onActivated()
+		}
+		
+		MenuItem{
+			//: 'Recordings' : Label for the recordings menu.
+			text: qsTr('recordings')
+			onTriggered: recordingsShortcut.onActivated()
 		}
 		
 		MenuItem {